Add files via upload
This code will get inside the DUDE molecule folder, and it will also get inside each protein folder. After that, it will extract
the value given by the 21st network (the DLS score network), and the value of the average given by the 20 networks from NNScore. FInally, it will write a file to store the 60 proteins' (in this case they were 60 readable proteins with ligands) scores.